+++ contract-testing/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased changes
+
+- Add methods to `Chain` for setting the exchange rates and block time based on queries from an external node.
+  - Add the method `setup_external_node_connection` for configuring the external node to use.
+  - Add the method `set_exchange_rates_via_external_node` for setting the exchange rates.
+  - Add the method `set_block_time_via_external_node` for setting the block time.
+
 ## 2.0.0
 
 - Include `ContractTraceElement`s for failed contract executions, including internal failures.